feat: Add chainIdName File
