+++ apps/web/src/hooks/useAccountEventListener.ts
@@ -2,7 +2,7 @@ import { useEffect, useMemo } from 'react'
 import { ExtendEthereum } from 'global'
 import replaceBrowserHistory from '@pancakeswap/utils/replaceBrowserHistory'
 import { ConnectorData } from 'wagmi'
-import { ChainIdName } from '@pancakeswap/sdk'
+import { ChainIdName } from 'config/ChainIdName'
 import { useAppDispatch } from '../state'
 import { clearUserStates } from '../utils/clearUserStates'
 import useActiveWeb3React from './useActiveWeb3React'
