+++ apps/web/src/hooks/useActiveChainId.ts
@@ -1,9 +1,10 @@
-import { ChainId, getChainId } from '@pancakeswap/sdk'
+import { ChainId } from '@pancakeswap/sdk'
 import { atom, useAtomValue } from 'jotai'
 import { useRouter } from 'next/router'
 import { useDeferredValue } from 'react'
 import { isChainSupported } from 'utils/wagmi'
 import { useNetwork } from 'wagmi'
+import { getChainId } from 'config/ChainIdName'
 import { useSessionChainId } from './useSessionChainId'
 
 const queryChainIdAtom = atom(-1) // -1 unload, 0 no chainId on query
