+++ apps/web/src/hooks/useActiveWeb3React.ts
@@ -4,7 +4,8 @@ import { useEffect, useRef } from 'react'
 import { EXCHANGE_PAGE_PATHS } from 'config/constants/exchange'
 import { isChainSupported } from 'utils/wagmi'
 import { useProvider } from 'wagmi'
-import { ChainId, getChainId, ChainIdName } from '@pancakeswap/sdk'
+import { ChainId } from '@pancakeswap/sdk'
+import { getChainId, ChainIdName } from 'config/ChainIdName'
 import { useActiveChainId } from './useActiveChainId'
 import { useSwitchNetworkLoading } from './useSwitchNetworkLoading'
